Changes to module packaging instructions and calls.
In webpack.config.js updated the output element with librarytarget
and library properties. This creates a DOM Window property with
the 'component package name'. The object associated with this property
has a 'default' property ( the default export) that is the ModuleComponents
class.  Bracket notation is then used on the Window object to get a new
instance of a ModuleComponent in CDTSInstructionsPromiseFactory.